chore: remove formService dependency  - move remaining logic to cardNetJobService
