+++ creditCard/src/main/java/com/kynsoft/finamer/creditcard/domain/services/ICardNetJobService.java
@@ -2,6 +2,10 @@ package com.kynsoft.finamer.creditcard.domain.services;
 
 import com.kynsoft.finamer.creditcard.domain.dto.CardnetJobDto;
 
-public interface ICardnetJobServeice {
+import java.util.UUID;
+
+public interface ICardNetJobService {
+    UUID create(CardnetJobDto dto);
+    CardnetJobDto findByTransactionId(UUID id);
     void update(CardnetJobDto dto);
 }
